Mu.GraphQL.Quasi

Description

Read .graphql files as a Schema and Package with one Service per object in the schema.

Documentation

graphql

Arguments

:: String
Name for the Package type, the Schema is derived from it

-> FilePath
Route to the file

-> Q [Dec]

Imports an GraphQL schema definition from a file.

type Primitives = [(Name, TypeQ)]

graphqlWithExtendedPrimitives

Arguments

:: Primitives

-> String
Name for the Package type, the Schema is derived from it

-> FilePath
Route to the file

-> Q [Dec]

Imports an GraphQL schema definition from a file.

graphql'

Arguments

:: Primitives

-> String
Name for the Schema type

-> String
Name for the Package type

-> FilePath
Route to the file

-> Q [Dec]

Imports an GraphQL schema definition from a file.