Mu.GraphQL.Query.Run

Documentation

runPipeline :: forall qr mut (p :: Package') pname ss hs chn qanns qmethods manns mmethods. (p ~ Package pname ss, LookupService ss qr ~ Service qr qanns qmethods, ParseMethod p qmethods, LookupService ss mut ~ Service mut manns mmethods, ParseMethod p mmethods, RunQueryFindHandler p hs chn ss (LookupService ss qr) hs, MappingRight chn qr ~ (), RunQueryFindHandler p hs chn ss (LookupService ss mut) hs, MappingRight chn mut ~ ()) => ServerT Identity chn p ServerErrorIO hs -> Proxy qr -> Proxy mut -> Maybe Text -> ExecutableDocument -> IO Value

runDocument :: (p ~ Package pname ss, RunQueryFindHandler p hs chn ss (LookupService ss qr) hs, MappingRight chn qr ~ (), RunQueryFindHandler p hs chn ss (LookupService ss mut) hs, MappingRight chn mut ~ ()) => ServerT Identity chn p ServerErrorIO hs -> Document p qr mut -> WriterT [GraphQLError] IO Value

runQuery :: forall p s pname ss hs sname sanns ms chn inh. (RunQueryFindHandler p hs chn ss s hs, p ~ Package pname ss, s ~ Service sname sanns ms, inh ~ MappingRight chn sname) => ServerT Identity chn p ServerErrorIO hs -> inh -> ServiceQuery p s -> WriterT [GraphQLError] IO Value