Mu.GraphQL.Server

Documentation

type GraphQLApp p qr mut sub m chn hs = (ParseTypedDoc p qr mut sub, RunDocument p qr mut sub m chn hs)

Run an GraphQL resolver directly

runGraphQLApp :: GraphQLApp p qr mut sub ServerErrorIO chn hs => Port -> ServerT chn p ServerErrorIO hs -> Proxy qr -> Proxy mut -> Proxy sub -> IO ()

Run a Mu graphQLApp on the given port.

runGraphQLAppSettings :: GraphQLApp p qr mut sub ServerErrorIO chn hs => Settings -> ServerT chn p ServerErrorIO hs -> Proxy qr -> Proxy mut -> Proxy sub -> IO ()

Run a Mu graphQLApp using the given Settings.

Go to Warp to declare Settings.

runGraphQLAppQuery :: GraphQLApp p ('Just qr) 'Nothing 'Nothing ServerErrorIO chn hs => Port -> ServerT chn p ServerErrorIO hs -> Proxy qr -> IO ()

Run a query-only Mu graphQLApp on the given port.

runGraphQLAppTrans :: GraphQLApp p qr mut sub m chn hs => Port -> (forall a. m a -> ServerErrorIO a) -> ServerT chn p m hs -> Proxy qr -> Proxy mut -> Proxy sub -> IO ()

Run a Mu graphQLApp on a transformer stack on the given port.

Build a WAI Application

graphQLApp :: GraphQLApp p qr mut sub ServerErrorIO chn hs => ServerT chn p ServerErrorIO hs -> Proxy qr -> Proxy mut -> Proxy sub -> Application

Turn a Mu GraphQL Server into a WAI Application.

These Applications can be later combined using, for example, wai-routes, or you can add middleware from wai-extra, among others.

graphQLAppTrans :: GraphQLApp p qr mut sub m chn hs => (forall a. m a -> ServerErrorIO a) -> ServerT chn p m hs -> Proxy qr -> Proxy mut -> Proxy sub -> Application

Query-only builders

graphQLAppQuery :: forall qr p chn hs. GraphQLApp p ('Just qr) 'Nothing 'Nothing ServerErrorIO chn hs => ServerT chn p ServerErrorIO hs -> Proxy qr -> Application

graphQLAppTransQuery :: forall qr m p chn hs. GraphQLApp p ('Just qr) 'Nothing 'Nothing m chn hs => (forall a. m a -> ServerErrorIO a) -> ServerT chn p m hs -> Proxy qr -> Application